Объединение тестов в один класс
