+++ src/test/java/CatTest.java
@@ -1,21 +1,36 @@
+import com.example.Cat;
+import com.example.Feline;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
-import com.example.Cat;
+
 import java.util.List;
+
 import static org.junit.Assert.assertTrue;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CatGetFoodTest {
+public class CatTest {
     private final List<String> expectedFoodList = List.of("Животные", "Птицы", "Рыба");
+    private final String expectedSound = "Мяу";
+
     @Mock
     Cat cat;
+    @Mock
+    Feline feline;
+
     @Test
     public void getFoodTest() throws Exception {
         Mockito.when(cat.getFood()).thenReturn(List.of("Животные", "Птицы", "Рыба"));
         assertTrue(expectedFoodList.containsAll(cat.getFood()) && cat.getFood().containsAll(expectedFoodList));
     }
+
+    @Test
+    public void getSoundTest() {
+        Cat cat = new Cat(feline);
+        assertTrue(expectedSound.equals(cat.getSound()));
+    }
+
 
 }
